# openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "RESOLA_TEST",
        "version": "0.1.0"
    },
    "paths": {
        "/api/v1/users/create": {
            "post": {
                "tags": [
                    "V1 Users"
                ],
                "summary": "Create User",
                "operationId": "Create_user_api_v1_users_create_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/users/update/{user_id}": {
            "put": {
                "tags": [
                    "V1 Users"
                ],
                "summary": "Update User",
                "operationId": "Update_user_api_v1_users_update__user_id__put",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "User Id",
                            "type": "integer"
                        },
                        "name": "user_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/users/delete/{user_id}": {
            "delete": {
                "tags": [
                    "V1 Users"
                ],
                "summary": "Delete User",
                "operationId": "Delete_user_api_v1_users_delete__user_id__delete",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "User Id",
                            "type": "integer"
                        },
                        "name": "user_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
        "/api/v1/books/create": {
            "post": {
                "tags": [
                    "V1 Books"
                ],
                "summary": "Create Book",
                "operationId": "Create_book_api_v1_books_create_post",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "Title",
                            "type": "string"
                        },
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Author",
                            "type": "string"
                        },
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Publish Date",
                            "type": "string",
                            "format": "date"
                        },
                        "name": "publish_date",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Isbn",
                            "type": "string"
                        },
                        "name": "isbn",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Price",
                            "type": "number"
                        },
                        "name": "price",
                        "in": "query"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_Create_book_api_v1_books_create_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/books/get/id/{book_id}": {
            "get": {
                "tags": [
                    "V1 Books"
                ],
                "summary": "Get Book By Id",
                "operationId": "Get_book_by_id_api_v1_books_get_id__book_id__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Book Id",
                            "type": "integer"
                        },
                        "name": "book_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/books/delete/{book_id}": {
            "delete": {
                "tags": [
                    "V1 Books"
                ],
                "summary": "Delete Book",
                "operationId": "Delete_book_api_v1_books_delete__book_id__delete",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Book Id",
                            "type": "integer"
                        },
                        "name": "book_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
        "/api/token": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "operationId": "login_api_token_post",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_login_api_token_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/signup": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Signup",
                "operationId": "signup_api_signup_post",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_signup_api_signup_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Body_Create_book_api_v1_books_create_post": {
                "title": "Body_Create_book_api_v1_books_create_post",
                "required": [
                    "file"
                ],
                "type": "object",
                "properties": {
                    "file": {
                        "title": "File",
                        "type": "string",
                        "format": "binary"
                    }
                }
            },
...
            "Body_login_api_token_post": {
                "title": "Body_login_api_token_post",
                "required": [
                    "username",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "grant_type": {
                        "title": "Grant Type",
                        "pattern": "password",
                        "type": "string"
                    },
                    "username": {
                        "title": "Username",
                        "type": "string"
                    },
                    "password": {
                        "title": "Password",
                        "type": "string"
                    },
                    "scope": {
                        "title": "Scope",
                        "type": "string",
                        "default": ""
                    },
                    "client_id": {
                        "title": "Client Id",
                        "type": "string"
                    },
                    "client_secret": {
                        "title": "Client Secret",
                        "type": "string"
                    }
                }
            },
            "Body_signup_api_signup_post": {
                "title": "Body_signup_api_signup_post",
                "required": [
                    "username",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "grant_type": {
                        "title": "Grant Type",
                        "pattern": "password",
                        "type": "string"
                    },
                    "username": {
                        "title": "Username",
                        "type": "string"
                    },
                    "password": {
                        "title": "Password",
                        "type": "string"
                    },
                    "scope": {
                        "title": "Scope",
                        "type": "string",
                        "default": ""
                    },
                    "client_id": {
                        "title": "Client Id",
                        "type": "string"
                    },
                    "client_secret": {
                        "title": "Client Secret",
                        "type": "string"
                    }
                }
            },
...
            "Book": {
                "title": "Book",
                "required": [
                    "image_link",
                    "id"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "title": "Title",
                        "type": "string"
                    },
                    "author": {
                        "title": "Author",
                        "type": "string"
                    },
                    "publish_date": {
                        "title": "Publish Date",
                        "type": "string",
                        "format": "date"
                    },
                    "isbn": {
                        "title": "Isbn",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "type": "number"
                    },
                    "image_link": {
                        "title": "Image Link",
                        "type": "string"
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    }
                }
            },
...
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "User": {
                "title": "User",
                "required": [
                    "id"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "title": "Email",
                        "type": "string"
                    },
                    "first_name": {
                        "title": "First Name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "type": "string"
                    },
                    "is_active": {
                        "title": "Is Active",
                        "type": "boolean",
                        "default": true
                    },
                    "is_superuser": {
                        "title": "Is Superuser",
                        "type": "boolean",
                        "default": false
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    }
                }
            },
            "UserCreate": {
                "title": "UserCreate",
                "required": [
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "title": "Email",
                        "type": "string"
                    },
                    "first_name": {
                        "title": "First Name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "type": "string"
                    },
                    "is_active": {
                        "title": "Is Active",
                        "type": "boolean",
                        "default": true
                    },
                    "is_superuser": {
                        "title": "Is Superuser",
                        "type": "boolean",
                        "default": false
                    },
                    "password": {
                        "title": "Password",
                        "type": "string"
                    }
                }
            },
            "UserUpdate": {
                "title": "UserUpdate",
                "type": "object",
                "properties": {
                    "email": {
                        "title": "Email",
                        "type": "string"
                    },
                    "first_name": {
                        "title": "First Name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "type": "string"
                    },
                    "is_active": {
                        "title": "Is Active",
                        "type": "boolean",
                        "default": true
                    },
                    "is_superuser": {
                        "title": "Is Superuser",
                        "type": "boolean",
                        "default": false
                    },
                    "password": {
                        "title": "Password",
                        "type": "string"
                    }
                }
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/token"
                    }
                }
            }
        }
    }
}